apps/backend/docker/paddleocr-service/requirements.txt
paddleocr==2.9.1
paddlepaddle==2.6.2
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
opencv-python-headless==4.10.0.84
pillow==11.0.0
numpy==1.26.4
